Add support for multiple input devices
Some 2-in-1s that come with a stylus will report the Stylus as a
separate input device (sometimes two separate devices) from the
Touchscreen. This adjusts the `-i/--touchscreen` argument to allow a
list of devices; such as the touchscreen, stylus, and eraser.

Still supports single touchscreens as well, i.e.:
- `-i "Wacom HID 4840 Finger"`
- `-i ""Wacom HID 4840 Finger" "Wacom HID 4840 Pen"`
- etc